search: remove alert proposal dependency on syntax package
