+++ cmd/frontend/graphqlbackend/search_alert_test.go
@@ -368,3 +368,23 @@ func TestAlertForOverRepoLimit(t *testing.T) {
 		})
 	}
 }
+
+func TestCapFirst(t *testing.T) {
+	tests := []struct {
+		name string
+		in   string
+		want string
+	}{
+		{name: "empty", in: "", want: ""},
+		{name: "a", in: "a", want: "A"},
+		{name: "ab", in: "ab", want: "Ab"},
+		{name: "хлеб", in: "хлеб", want: "Хлеб"},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := capFirst(tt.in); got != tt.want {
+				t.Errorf("makeTitle() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
